Segunda parte, ingresos, salidas, reservas

+++ database/migrations/2023_05_13_220113_create_Guardias_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('Guardias', function (Blueprint $table) {
             $table->integer('id_guardia')->primary();
             $table->date('fecha_incorporacion')->nullable();
-            $table->integer('id_usuario')->nullable()->index('fk_guardia_usuario_idx');
+            $table->unsignedBigInteger('id_usuario')->nullable()->index('fk_guardia_usuario_idx');
         });
     }
